ann_yoddha_backend/pyproject.toml
[project]
name = "ann-yoddha-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.47",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
